Fix changeColor() modifying the original color object
Due to an oversight in the color transition code, calling changeColor()
modified the original color object passed to the HPGauge constructor.
This isn't a problem if the gauge is initialized using a predefined
color, since those are created anew on each access.  The showcase on the
other hand cycles through an array of colors, which triggered the bug.